% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_plotly.R
\name{alluvial_as_plotly-shiny}
\alias{alluvial_as_plotly-shiny}
\alias{alluvial_as_plotlyOutput}
\alias{render_alluvial_as_plotly}
\title{Shiny bindings for alluvial_as_plotly}
\usage{
alluvial_as_plotlyOutput(outputId, width = "100\%", height = "100\%",
  inline = F)

render_alluvial_as_plotly(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a alluvial_as_plotly}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using alluvial_as_plotly within Shiny
applications and interactive Rmd documents.
}
