% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial_model_response.R
\name{alluvial_model_response}
\alias{alluvial_model_response}
\title{FUNCTION_TITLE}
\usage{
alluvial_model_response(pred, dspace, imp, degree = 3, bins = 5,
  force = FALSE)
}
\arguments{
\item{pred}{PARAM_DESCRIPTION}

\item{dspace}{PARAM_DESCRIPTION}

\item{imp}{PARAM_DESCRIPTION}

\item{degree}{PARAM_DESCRIPTION, Default: 3}

\item{bins}{PARAM_DESCRIPTION, Default: 5}

\item{force}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[easyalluvial]{alluvial_wide}}
}
