% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial_model_response.R
\name{alluvial_model_response_caret}
\alias{alluvial_model_response_caret}
\title{FUNCTION_TITLE}
\usage{
alluvial_model_response_caret(train, degree = 3, bins = 5, force = F)
}
\arguments{
\item{train}{PARAM_DESCRIPTION}

\item{degree}{PARAM_DESCRIPTION, Default: 3}

\item{bins}{PARAM_DESCRIPTION, Default: 5}

\item{force}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[caret]{varImp}},\code{\link[caret]{extractPrediction}}
}
