% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial_model_response.R
\name{check_pkg_installed}
\alias{check_pkg_installed}
\title{check if package is installed}
\usage{
check_pkg_installed(pkg)
}
\arguments{
\item{pkg}{character, package name}
}
\value{
logical
}
\description{
check if package is installed
}
\examples{
check_pkg_installed("easyalluvial")

}
