% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_manip.R
\name{f_manip_bin_numerics}
\alias{f_manip_bin_numerics}
\title{bin numerical columns}
\usage{
f_manip_bin_numerics(df, bins = 5, bin_labels = c("LL", "ML", "M", "MH",
  "HH"), center = T, scale = T, transform = T)
}
\arguments{
\item{df}{dataframe with numeric variables}

\item{bins}{number of bins for numerical variables, Default: 5}

\item{bin_labels}{labels for the bins from low to high, Default: c("LL",
"ML", "M", "MH", "HH")#' @param center boolean, Default: T}

\item{center}{boolean, Default: T}

\item{scale}{boolean, Default: T}

\item{transform}{boolean, Default: T}
}
\value{
dataframe
}
\description{
centers, scales and Yeo Johnson transforms numeric variables in
  a dataframe before binning into n bins of eqal range. Outliers based on
  boxplot stats are capped (set to min or max of boxplot stats).
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
