% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_manip.R
\name{f_manip_factor_2_numeric}
\alias{f_manip_factor_2_numeric}
\title{converts factor to numeric preserving numeric levels and order in character levels}
\usage{
f_manip_factor_2_numeric(vec)
}
\arguments{
\item{vec}{vector}
}
\value{
vector
}
\examples{
fac_num = factor( c(1,3,8) )
fac_chr = factor( c('foo','bar') )
fac_chr_ordered = factor( c('a','b','c'), ordered = T )

f_manip_factor_2_numeric( fac_num )
f_manip_factor_2_numeric( fac_chr )
f_manip_factor_2_numeric( fac_chr_ordered )
}
\seealso{
\code{\link[stringr]{str_detect}}
}
