% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot.R
\name{f_plot_adjust_col_vector_length}
\alias{f_plot_adjust_col_vector_length}
\title{adjust length of color vector, by repeating colors}
\usage{
f_plot_adjust_col_vector_length(n = 74, col_vector = f_plot_col_vector74())
}
\arguments{
\item{n}{length, Default: 74}

\item{col_vector}{vector containing colors, Default: f_plot_col_vector74()}
}
\value{
vector containing colors of specified length
}
\examples{

length( f_plot_adjust_col_vector_length(100) )

}
