% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial_wide.R
\name{f_plot_alluvial}
\alias{f_plot_alluvial}
\title{plot alluvial on tidy data}
\usage{
f_plot_alluvial(data, variables = names(data), id = NULL,
  max_variables = 20, bins = 5, bin_labels = c("LL", "ML", "M", "MH",
  "HH"), NA_label = "NA", order_levels = NULL, fill_by = "first_variable",
  col_vector_flow = f_plot_col_vector74(faint = F, greys = F),
  col_vector_value = RColorBrewer::brewer.pal(9, "Greys")[c(3, 6, 4, 7, 5)],
  verbose = F)
}
\arguments{
\item{data}{a dataframe}

\item{variables}{vector denoting names and order of the plotted variables,
Default: names(data)}

\item{id}{character vector denoting id column}

\item{max_variables}{maximum number of variables, Default: 20}

\item{bins}{number of bins for numerical variables, Default: 5}

\item{bin_labels}{labels for the bins from low to high, Default: c("LL",
"ML", "M", "MH", "HH")}

\item{NA_label}{character vector define label for missing data}

\item{order_levels}{character vector denoting levels to be reorderer from low to high}

\item{fill_by}{one_of(c('first_variable', 'last_variable', 'all_flows',
'values')), Default: 'first_variable'}

\item{col_vector_flow}{HEX colors for flows, Default:
f_plot_col_vector74(faint = F, greys = F)}

\item{col_vector_value}{Hex colors for y levels/values, Default:
RColorBrewer::brewer.pal(9, "Greys")[c(3, 6, 4, 7, 5)]}

\item{verbose}{logical, print plot summary, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
plots a dataframe as an alluvial plot. All numerical variables
  are scaled, centered and YeoJohnson transformed before binning.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){

data_ls = mtcars \%>\%
  f_clean_data()

data = data_ls$data
max_variables = 5
variables = c( data_ls$categoricals[1:3], data_ls$numericals[1:3] )

f_plot_alluvial( data = data
                , variables = variables
                , max_variables = max_variables
                , fill_by = 'first_variable' )

f_plot_alluvial( data = data
                , variables = variables
                , max_variables = max_variables
                , fill_by = 'last_variable' )

f_plot_alluvial( data = data
                , variables = variables
                , max_variables = max_variables
                , fill_by = 'all_flows' )

f_plot_alluvial( data = data
                , variables = variables
                , max_variables = max_variables
                , fill_by = 'first_variable' )

# manually order variable values

f_plot_alluvial( data = data
                 , variables = variables
                 , max_variables = max_variables
                 , fill_by = 'values'
                 , order_levels = c('1', '0') )
}
}
}
\seealso{
\code{\link[RColorBrewer]{brewer.pal}}
  \code{\link[forcats]{fct_relevel}}
  \code{\link[ggalluvial]{geom_flow}},\code{\link[ggalluvial]{geom_stratum}}
}
