% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial_model_response.R
\name{get_pdp_predictions}
\alias{get_pdp_predictions}
\title{get predictions compatibel with the partial dependency plotting method}
\usage{
get_pdp_predictions(df, imp, .f_predict, m, degree = 4, bins = 5)
}
\arguments{
\item{df}{dataframe, training data}

\item{imp}{dataframe, with not more then two columns one of them numeric
containing importance measures and one character or factor column containing
corresponding variable names as found in training data.}

\item{.f_predict}{corresponding model predict() function. Often predict
functions are undocumented and can be found using `:::`. For example
`randomForest:::predict.randomForest`. Predictict functions needs to accept
`m` as the first parameter and use the `newdata` parameter. Supply a wrapper
for predict functions with x-y synthax.}

\item{m}{model object}

\item{degree}{integer,  number of top important variables to select. For
plotting more than 4 will result in two many flows and the alluvial plot
will not be very readable, Default: 4}

\item{bins}{integer, number of bins for numeric variables, increasing this
number might result in too many flows, Default: 5}
}
\value{
vector, predictions
}
\description{
the partial dependency plotting method uses the averaged
 predictions of all observations in the training data in which the values of
 the variables of interest have been modified. The variables of interest
 being those for which we want to plot the model response. This is a wrapper
 that iterates over all rows in the training data and calls
 \code{\link[easyalluvial]{get_data_space}}, feeds it into the supplied
 predict function and averages the results.
}
\details{
DETAILS
}
\examples{
 df = mtcars2[, ! names(mtcars2) \%in\% 'ids' ]
 m = randomForest::randomForest( disp ~ ., df)
 imp = m$importance

 pred = get_pdp_predictions(df, imp
                            , .f_predict = randomForest:::predict.randomForest
                            , m
                            , degree = 3
                            , bins = 5)

}
\seealso{
\code{\link[progress]{progress_bar}}
}
