% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.R
\name{manip_explicit_na}
\alias{manip_explicit_na}
\title{Make NA levels explicit}
\usage{
manip_explicit_na(f, na_level)
}
\description{
fct_na_value_to_level will create a NA level even there is
none present to begin with
}
\keyword{internal}
