% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.R
\name{manip_get_ggplot_data}
\alias{manip_get_ggplot_data}
\title{Get ggplot data}
\usage{
manip_get_ggplot_data(p)
}
\arguments{
\item{p}{ggplot object}
}
\description{
retrieve ggplot 2 plot data from S3 and S7 (>= ggplot2 v4.0)
}
