% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{palette_plot_rgp}
\alias{palette_plot_rgp}
\title{plot rgb values of palette}
\usage{
palette_plot_rgp(palette)
}
\arguments{
\item{palette}{any vector containing color hex values}
}
\value{
ggplot2 plot
}
\examples{
\dontrun{
if(interactive()){
palette_qualitative() \%>\%
  palette_filter( thresh = 50) \%>\%
  palette_plot_rgp()
 }
}
}
\seealso{
\code{\link[easyaluvial]{palette_plot_intensity}}
}
