% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{palette_qualitative}
\alias{palette_qualitative}
\title{compose palette from qualitative RColorBrewer palettes}
\usage{
palette_qualitative()
}
\value{
vector with hex values
}
\description{
combines all unique values found in all qualitative RColorBrewer palettes
}
\examples{
palette_qualitative()
}
\seealso{
\code{\link[RColorBrewer]{RColorBrewer}}
}
