% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial_model_response.R
\name{use_e1071}
\alias{use_e1071}
\title{calls e1071::skewness}
\usage{
use_e1071(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
if e1071 is not listed a a dependency I get an error. I assume
  caret uses it to calculate feature importance. However, e1071 is not listed
  as a caret dependency. I have to add a function that directly calls it, so
  I do not get a NOTE from RMD Check on Linux.
}
\seealso{
\code{\link[e1071]{skewness}}
}
