% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial_wide.R
\name{alluvial_wide}
\alias{alluvial_wide}
\title{alluvial plot of data in wide format}
\usage{
alluvial_wide(data, id = NULL, max_variables = 20, bins = 5,
  bin_labels = c("LL", "ML", "M", "MH", "HH"), NA_label = "NA",
  order_levels = NULL, fill_by = "first_variable",
  col_vector_flow = palette_qualitative() \%>\% palette_filter(greys =
  F), col_vector_value = RColorBrewer::brewer.pal(9, "Greys")[c(3, 6, 4,
  7, 5)], verbose = F)
}
\arguments{
\item{data}{a dataframe}

\item{id}{unquoted column name of id column}

\item{max_variables}{maximum number of variables, Default: 20}

\item{bins}{number of bins for numerical variables, Default: 5}

\item{bin_labels}{labels for the bins from low to high, Default: c("LL",
"ML", "M", "MH", "HH")}

\item{NA_label}{character vector define label for missing data, Default: 'NA'}

\item{order_levels}{character vector denoting levels to be reorderer from low to high}

\item{fill_by}{one_of(c('first_variable', 'last_variable', 'all_flows',
'values')), Default: 'first_variable'}

\item{col_vector_flow}{HEX colors for flows, Default: palette_filter( greys = F)}

\item{col_vector_value}{Hex colors for y levels/values, Default:
RColorBrewer::brewer.pal(9, "Greys")[c(3, 6, 4, 7, 5)]}

\item{verbose}{logical, print plot summary, Default: F}
}
\value{
alluvial plot
}
\description{
plots a dataframe as an alluvial plot. All numerical variables
  are scaled, centered and YeoJohnson transformed before binning. Plots all
  variables in the sequence as they appear in the dataframe until maximum
  number of values is reached.
}
\examples{
\dontrun{
if(interactive()){

require(tidyverse)

data = as_tibble(mtcars)
categoricals = c('cyl', 'vs', 'am', 'gear', 'carb')
numericals = c('mpg', 'cyl', 'disp', 'hp', 'drat', 'wt', 'qsec')
max_variables = 5

data = data \%>\%
  mutate_at( vars(categoricals), as.factor )


alluvial_wide( data = data
                , max_variables = max_variables
                , fill_by = 'first_variable' )

alluvial_wide( data = data
                , max_variables = max_variables
                , fill_by = 'last_variable' )

alluvial_wide( data = data
                , max_variables = max_variables
                , fill_by = 'all_flows' )

alluvial_wide( data = data
                , max_variables = max_variables
                , fill_by = 'first_variable' )

# manually order variable values

alluvial_wide( data = data
                 , max_variables = max_variables
                 , fill_by = 'values'
                 , order_levels = c('1', '0') )
}
}
}
\seealso{
\code{\link[easyalluvial]{alluvial_wide}}
  \code{\link[ggalluvial]{geom_flow}},\code{\link[ggalluvial]{geom_stratum}}
}
