% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial_model_response.R
\name{get_data_space}
\alias{get_data_space}
\title{FUNCTION_TITLE}
\usage{
get_data_space(df, imp, degree = 3, bins = 5)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{imp}{PARAM_DESCRIPTION}

\item{degree}{PARAM_DESCRIPTION, Default: 3}

\item{bins}{PARAM_DESCRIPTION, Default: 5}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
